Replace protoList with protoIDtoPD map, unify protocol id namespace
In Setup.Execute, there were two data structures: a protocol list, and a
protocol map. The protocol list was merely used to generate the protocol
map. Therefore, it was removed, and the code was refactored to directly
generate a protocol map.

In the protocol map generation, protocol ids were hardcoded strings.
This was changed such that a protocol ID is always the string
representation of the protocol type. For protocols
of the type Rotation Key Generation (RTG), the string
representation is in the form "RTG[<galoisElement>]". This is because a
setup phase can generate more than one rotation key, hence we need a way
to uniquely identify each of them.
The same convention is used to save and store keys to and from the
ObjectStore.