Add compute phase to node application
Added the computational phase to the main node application (./apps/node).
Currently, the app executes a test circuit (identity + CKS) with fixed
aggregator and receivers. In particular, the aggregator has ID "cloud",
and the only party providing input has ID "node-a", this party is also
the only receiver. In the future, it will be possible to specify the details
of the computational phase using a compute descriptor.

To maintain backward compatibility and leave unchanged the current
scenarios, it is possible to specify whether the node should execute the
compute phase with the flag -docompute (disabled by default).