Add full PCKS integration in the setup and compute phase
Now the setup phase can be used to send the public key of a receiver
node to the other parties. The party that sends its public key generates
it on the fly during the setup phase. The key is sent using a new PAT
protocol: protocols.PK. This protocol is almost identical to the
protocols.CKG protocol, except for the generation of the secret key.
In the future, this new protocol can be refactored to avoid code
duplication.