Fix setup descr. parsing, augument localtest, cleanup
The flag -docompute is true by default. If the flag is set to false, and
a JSON setup description is provided, then the setup description is
parsed from a file, instead of being inferred from the circuit
description.

The localtest for the Public Key sending now has three session nodes.

Removed unused files.