Unify naming for output keypair
Added functions to store and load public and private keys used to
encrypt and decrypt the output. The relative errors are now handled
better. The string used to index the ObjectStore is only present in the
new function, as opposed to multiple files, which make the code less
error-prone and more readable.