Add getters and setters for session keys. Fix PK indexing
Now HElium abstracts the access to the keys in the ObjectStore using
getters and setters in the Session. This way, the user does not have to
craft strings to index the keys inside the ObjectStore, the user can
just use the methods provided.

This also fixes a bug in the indexing of public keys shared by external
receivers.

The keys inside the ObjectStore are now indexed using the
`protocols.Signature.ToObjectStore` function, that returns a simplified string
representation of a protocol signature.

This refactor is retrocompatible with the Setup.Execute function, that
saves generated keys accessing directly the ObjectStore and using signatures.

Some code redundancy is due by the fact that it is not possible to
use the methods of the package `protocols` inside the package `session`
because this would create a circular dependency.